Timeline added.
Use this version to start integration and testing of whole system. This version has the following issues:
-E-mail verification not implemented
-Non-MUN emails not blocked

However, you can start adding and combing your modules so that they all link together and work once a user logs in. Since my modules are fairly independent of yours, we can use this for now. I'll work on fixing the missing features and adding them in the meantime since integration can proceed without them. If needed, we may have to demonstrate without those features.

For integration here are some basics of what we must do:
-Added code snippets to the timeline to tell handlebars where returned values from DB go. Look to other handlebar pages for an understanding.
-Create your own handlebar pages for modules that require a separate page. Any modules that are on the same page as another, simply edit that handlebars file to include the HTML for the module. Don't forget to create a separate CSS file for your HTML and l